+++ packages/marshal/test/test-marshal-smallcaps.js
@@ -160,6 +160,10 @@ test('smallcaps unserialize errors', t => {
 });
 
 test('smallcaps unserialize extended errors', t => {
+  if (typeof AggregateError === 'undefined') {
+    t.pass('skip test on platforms prior to AggregateError');
+    return;
+  }
   const { unserialize } = makeTestMarshal();
   const uns = body => unserialize({ body, slots: [] });
 
@@ -170,7 +174,6 @@ test('smallcaps unserialize extended errors', t => {
   t.false('extraProp' in refErr);
   t.false('cause' in refErr);
   t.false('errors' in refErr);
-  console.log('error with extra prop', refErr);
 
   const aggErr = uns(
     '#{"#error":"msg","name":"AggregateError","extraProp":"foo","cause":"bar","errors":["zip","zap"]}',
@@ -179,7 +182,6 @@ test('smallcaps unserialize extended errors', t => {
   t.false('extraProp' in aggErr);
   t.false('cause' in aggErr);
   t.is(aggErr.errors.length, 0);
-  console.log('error with extra prop', aggErr);
 
   const unkErr = uns(
     '#{"#error":"msg","name":"UnknownError","extraProp":"foo","cause":"bar","errors":["zip","zap"]}',
@@ -188,10 +190,13 @@ test('smallcaps unserialize extended errors', t => {
   t.false('extraProp' in unkErr);
   t.false('cause' in unkErr);
   t.false('errors' in unkErr);
-  console.log('error with extra prop', unkErr);
 });
 
 test('smallcaps unserialize errors w recognized extensions', t => {
+  if (typeof AggregateError === 'undefined') {
+    t.pass('skip test on platforms prior to AggregateError');
+    return;
+  }
   const { unserialize } = makeTestMarshal();
   const uns = body => unserialize({ body, slots: [] });
 
@@ -204,7 +209,6 @@ test('smallcaps unserialize errors w recognized extensions', t => {
   t.false('extraProp' in refErr);
   t.is(getPrototypeOf(refErr.cause), URIError.prototype);
   t.is(getPrototypeOf(refErr.errors[0]), URIError.prototype);
-  console.log('error with extra prop', refErr);
 
   const aggErr = uns(
     `#{"#error":"msg","name":"AggregateError","extraProp":"foo","cause":${errEnc},"errors":[${errEnc}]}`,
@@ -213,7 +217,6 @@ test('smallcaps unserialize errors w recognized extensions', t => {
   t.false('extraProp' in aggErr);
   t.is(getPrototypeOf(refErr.cause), URIError.prototype);
   t.is(getPrototypeOf(refErr.errors[0]), URIError.prototype);
-  console.log('error with extra prop', aggErr);
 
   const unkErr = uns(
     `#{"#error":"msg","name":"UnknownError","extraProp":"foo","cause":${errEnc},"errors":[${errEnc}]}`,
@@ -222,7 +225,6 @@ test('smallcaps unserialize errors w recognized extensions', t => {
   t.false('extraProp' in unkErr);
   t.is(getPrototypeOf(refErr.cause), URIError.prototype);
   t.is(getPrototypeOf(refErr.errors[0]), URIError.prototype);
-  console.log('error with extra prop', unkErr);
 });
 
 test('smallcaps mal-formed @qclass', t => {