+++ packages/ses/test/error/test-aggregate-error-console.js
@@ -5,6 +5,10 @@ import { throwsAndLogs } from './throws-and-logs.js';
 lockdown();
 
 test('aggregate error console', t => {
+  if (typeof AggregateError === 'undefined') {
+    t.pass('skip test on platforms prior to AggregateError');
+    return;
+  }
   const e3 = Error('e3');
   const e2 = Error('e2', { cause: e3 });
   const u4 = URIError('u4', { cause: e2 });