+++ packages/ses/test/error/test-aggregate-error.js
@@ -6,6 +6,10 @@ const { getOwnPropertyDescriptor } = Object;
 lockdown();
 
 test('aggregate error', t => {
+  if (typeof AggregateError === 'undefined') {
+    t.pass('skip test on platforms prior to AggregateError');
+    return;
+  }
   const e1 = Error('e1');
   const e2 = Error('e2', { cause: e1 });
   const u3 = URIError('u3', { cause: e1 });
@@ -28,6 +32,10 @@ test('aggregate error', t => {
 });
 
 test('Promise.any aggregate error', async t => {
+  if (typeof AggregateError === 'undefined') {
+    t.pass('skip test on platforms prior to AggregateError');
+    return;
+  }
   const e1 = Error('e1');
   const e2 = Error('e2', { cause: e1 });
   const u3 = URIError('u3', { cause: e1 });